pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "openapi-cli-gen"
version = "0.0.1"
description = "Generate typed Python CLIs from OpenAPI specs with Pydantic model flattening into CLI flags"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "shivaam" },
]
keywords = ["openapi", "cli", "pydantic", "fastapi", "code-generator", "typer"]
classifiers = [
    "Development Status :: 1 - Planning",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Code Generators",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "pydantic-settings>=2.13",
    "jsonref>=1.1",
    "openapi-pydantic>=0.5",
    "httpx>=0.27",
    "pyyaml>=6.0",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "datamodel-code-generator>=0.56",
    "ruff>=0.4",
    "fastapi>=0.115",
    "uvicorn>=0.30",
    "typer>=0.12",
    "pydanclick>=0.5",
]

[project.urls]
Homepage = "https://github.com/shivaam/openapi-cli-gen"
Repository = "https://github.com/shivaam/openapi-cli-gen"
Issues = "https://github.com/shivaam/openapi-cli-gen/issues"
